this fixes #967 this fixes #955
In file clim-core/commands.lisp

Add an :initform nil for the docuemtation field of %menu-item.  This
is necessary to avoid an error if the mouse moves over a command menu
item.  The problem is that if a command-menu-item takes arguments, when
the mouse moves over the item an error will signaled rather than
querying for the rest of the arguments; this is because the
documentation was not being initialized.

In file clim-core/gadgets/menu.lisp

Introduce a new presentation type command-menu-item to be used in command menus diplayed in appllication-frames.
Add a translator from the command-menu-item to command.
Use the new presentation-type when displaying command table menu

In the ACCEPT presentation-method for COMMAND:

1. Use WITH-INPUT-CONTEXT for presentation-type 'command so that this method
    can handle mouse clicks on objects that translate to (or are presented as) commands

2. In the handler for WITH-INPUT-CONTEXT, check if the command is a partial command and
    if so call the partial command parser and note that the input needs to be replaced

3. If the input needs to be replaced, then explictly provide the :RESCAN NIL option to
    PRESENTATION-REPLACE-INPUT.
    This has a different effect than not providing :rescan at all (which seems to default to T).

The reason for 3 is that the command returned by clicking on an object may contain arguments
 that print (or present) unreadably. PRESENTATION-REPLACE-INPUT presents that command into the input
 buffer and if :RESCAN NIL is not provided, then ACCEPT-1 will call the ACCEPT presentation method,
 invoking the command-parser again.  However, this time the input buffer contains the unreadable string
 for the argument in the command and the command argument parser will then signal an error.
 This was observed in Clouseau.

The reason for 2 is that a command-menu in an application might contain a command with missing arguments.
 Without this fix, clicking on the item will cause and error message saying that the command is incomplete.